Product Book, Payment, Admin View/Edit Booked Items

+++ cart/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
-from .models import Cart, CartItem, Product
+from .models import Cart, CartItem, Product, Orders
 from accounts import forms, models
 
 # Create your views here.
@@ -19,9 +19,6 @@ def add_cart(request, product_id):
     product = Product.objects.get(id=product_id)
 
     if request.method == "POST":
-        # for item in request.POST:
-        #     key = item
-        #     value = request.POST[key]
 
         product = Product.objects.get(id=product_id)
         try:
@@ -73,9 +70,31 @@ def remove_cart_item(request, cart_item_id):
     messages.success(request, "Item Sucessfully Removed")
     return redirect('cart')
 
+def payment(request):
+    return render(request, 'store/payment.html')
 
-def purchaseitem(request, cart_item_id):
-    cart_item = CartItem.objects.get(id=cart_item_id)
-    cart_item.delete()
-    messages.success(request, "Item Sucessfully Bought")
-    return redirect('cart')
+def completeOrder(request):
+    return render(request, 'store/ordercomplete.html')
+
+# def purchaseitem(request, cart_item_id):
+#     cart_item = CartItem.objects.get(id=cart_item_id)
+#     cart_item.delete()
+#     messages.success(request, "Item Sucessfully Bought")
+#     return redirect('cart')
+
+def purchaseitem(request, product_id):
+    if request.method == "POST":
+        current_user = request.user
+        product = Product.objects.get(id=product_id)
+
+
+
+        order = Orders(user=current_user, product=product)
+        order.save()
+
+        cart_item_id  = request.POST['cart_item_id']
+        cart_item = CartItem.objects.get(id=cart_item_id)
+        cart_item.delete()
+
+        messages.success(request, "Item Ordered")
+        return redirect('payment')
